llm-chat-indexer/requirements.txt
# Web Framework
Flask>=2.0.0
Werkzeug>=2.0.0
flask-swagger-ui>=4.11.1
apispec>=6.0.0

# Vector Database and Embeddings
chromadb>=0.4.22
sentence-transformers>=2.2.0

# Natural Language Processing
spacy>=3.5.0

# Knowledge Graph
networkx>=3.0

# LLM Integration
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-community>=0.0.10

# Configuration and Environment
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.1.0

# Development Tools
flake8>=6.0.0
black>=23.0.0